Start of test fixing Converted tests to xunit, joined both projects since they both only test the lib anyway. They still fail, but that's to be fixed later. I just wanted start organizing it for now.
