+++ UndertaleModLib.Tests/GameLoadingTests.cs
@@ -1,15 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics;
-using System.IO;
-using System.Security.Cryptography;
+using System.Linq;
+using System.Text;
 using System.Threading.Tasks;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using UndertaleModLib;
+//using UndertaleModLib;
 using UndertaleModLib.Decompiler;
 using UndertaleModLib.Models;
 
-namespace UndertaleModTests
+namespace UndertaleModLib.Tests
 {
     public abstract class GameLoadingTestBase : GameTestBase
     {
@@ -17,7 +15,7 @@ namespace UndertaleModTests
         {
         }
 
-        [TestMethod]
+        [Fact]
         public void SaveDataAndCompare()
         {
             using (MemoryStream ms = new MemoryStream())
@@ -26,11 +24,11 @@ namespace UndertaleModTests
 
                 ms.Position = 0;
                 string writtenMD5 = GenerateMD5(ms);
-                Assert.AreEqual(expectedMD5, writtenMD5, "Written file doesn't match read file");
+                Assert.True(expectedMD5 == writtenMD5, "Written file doesn't match read file");
             }
         }
 
-        [TestMethod]
+        [Fact]
         public void DecompileAllScripts()
         {
             GlobalDecompileContext context = new GlobalDecompileContext(data, true);
@@ -39,7 +37,7 @@ namespace UndertaleModTests
                 //Console.WriteLine(code.Name.Content);
                 try
                 {
-                    Decompiler.Decompile(code, context);
+                    Decompiler.Decompiler.Decompile(code, context);
                 }
                 catch (Exception e)
                 {
@@ -48,7 +46,7 @@ namespace UndertaleModTests
             });
         }
 
-        [TestMethod]
+        [Fact]
         public void DisassembleAndReassembleAllScripts()
         {
             Parallel.ForEach(data.Code, (code) =>
@@ -56,7 +54,7 @@ namespace UndertaleModTests
                 //Console.WriteLine(code.Name.Content);
 
                 bool knownBug = false;
-                foreach(var instr in code.Instructions)
+                foreach (var instr in code.Instructions)
                 {
                     if (instr.Value?.GetType() == typeof(UndertaleResourceById<UndertaleString, UndertaleChunkSTRG>))
                     {
@@ -82,43 +80,50 @@ namespace UndertaleModTests
                 }
 
                 IList<UndertaleInstruction> reasm = Assembler.Assemble(disasm, data.Functions, data.Variables, data.Strings);
-                Assert.AreEqual(code.Instructions.Count, reasm.Count, "Reassembled instruction count didn't match the disassembly for script " + code.Name.Content);
-                for(int i = 0; i < code.Instructions.Count; i++)
+
+                Assert.True(code.Instructions.Count == reasm.Count, "Reassembled instruction count didn't match the disassembly for script " + code.Name.Content);
+
+                for (int i = 0; i < code.Instructions.Count; i++)
                 {
                     string errMsg = "Instruction at " + code.Instructions[i].Address.ToString("D5") + " didn't match for script: " + code.Name.Content;
-                    Assert.AreEqual(code.Instructions[i].Kind, reasm[i].Kind, errMsg);
-                    Assert.AreEqual(code.Instructions[i].ComparisonKind, reasm[i].ComparisonKind, errMsg);
-                    Assert.AreEqual(code.Instructions[i].Type1, reasm[i].Type1, errMsg);
-                    Assert.AreEqual(code.Instructions[i].Type2, reasm[i].Type2, errMsg);
-                    Assert.AreEqual(code.Instructions[i].TypeInst, reasm[i].TypeInst, errMsg);
-                    Assert.AreEqual(code.Instructions[i].Extra, reasm[i].Extra, errMsg);
-                    Assert.AreEqual(code.Instructions[i].SwapExtra, reasm[i].SwapExtra, errMsg);
-                    Assert.AreEqual(code.Instructions[i].ArgumentsCount, reasm[i].ArgumentsCount, errMsg);
-                    Assert.AreEqual(code.Instructions[i].JumpOffsetPopenvExitMagic, reasm[i].JumpOffsetPopenvExitMagic, errMsg);
-                    if (!code.Instructions[i].JumpOffsetPopenvExitMagic)
-                        Assert.AreEqual(code.Instructions[i].JumpOffset, reasm[i].JumpOffset, errMsg); // note: also handles IntArgument implicitly
-                    Assert.AreSame(code.Instructions[i].Destination?.Target, reasm[i].Destination?.Target, errMsg);
-                    Assert.AreEqual(code.Instructions[i].Destination?.Type, reasm[i].Destination?.Type, errMsg);
-                    Assert.AreSame(code.Instructions[i].Function?.Target, reasm[i].Function?.Target, errMsg);
-                    Assert.AreEqual(code.Instructions[i].Function?.Type, reasm[i].Function?.Type, errMsg);
 
-                    Assert.AreEqual(code.Instructions[i].Value?.GetType(), reasm[i].Value?.GetType(), errMsg);
+                    Assert.True(code.Instructions[i].Kind == reasm[i].Kind, errMsg);
+                    Assert.True(code.Instructions[i].ComparisonKind == reasm[i].ComparisonKind, errMsg);
+                    Assert.True(code.Instructions[i].Type1 == reasm[i].Type1, errMsg);
+                    Assert.True(code.Instructions[i].Type2 == reasm[i].Type2, errMsg);
+                    Assert.True(code.Instructions[i].TypeInst == reasm[i].TypeInst, errMsg);
+                    Assert.True(code.Instructions[i].Extra == reasm[i].Extra, errMsg);
+                    Assert.True(code.Instructions[i].SwapExtra == reasm[i].SwapExtra, errMsg);
+                    Assert.True(code.Instructions[i].ArgumentsCount == reasm[i].ArgumentsCount, errMsg);
+                    Assert.True(code.Instructions[i].JumpOffsetPopenvExitMagic == reasm[i].JumpOffsetPopenvExitMagic, errMsg);
+
+                    if (!code.Instructions[i].JumpOffsetPopenvExitMagic)
+                        Assert.True(code.Instructions[i].JumpOffset == reasm[i].JumpOffset, errMsg); // note: also handles IntArgument implicitly
+
+                    Assert.True(Object.ReferenceEquals(code.Instructions[i].Destination?.Target, reasm[i].Destination?.Target), errMsg);
+                    Assert.True(code.Instructions[i].Destination?.Type == reasm[i].Destination?.Type, errMsg);
+                    Assert.True(Object.ReferenceEquals(code.Instructions[i].Function?.Target, reasm[i].Function?.Target), errMsg);
+                    Assert.True(code.Instructions[i].Function?.Type == reasm[i].Function?.Type, errMsg);
+
+                    Assert.True(Object.ReferenceEquals(code.Instructions[i].Value?.GetType(), reasm[i].Value?.GetType()), errMsg);
+
                     if (code.Instructions[i].Value?.GetType() == typeof(double))
-                        Assert.AreEqual((double)code.Instructions[i].Value, (double)reasm[i].Value, Math.Abs((double)code.Instructions[i].Value) * (1e-5), errMsg); // see issue #53
+                        Assert.True(Math.Abs((double)code.Instructions[i].Value - (double)reasm[i].Value)
+                            <= (Math.Abs((double)code.Instructions[i].Value) * (1e-5)), errMsg); // see issue #53
                     else if (code.Instructions[i].Value?.GetType() == typeof(float))
-                        Assert.AreEqual((float)code.Instructions[i].Value, (float)reasm[i].Value, Math.Abs((float)code.Instructions[i].Value) * (1e-5), errMsg); // see issue #53
+                        Assert.True(Math.Abs((float)code.Instructions[i].Value - (float)reasm[i].Value)
+                            <= (Math.Abs((float)code.Instructions[i].Value) * (1e-5)), errMsg); // see issue #53
                     else if (code.Instructions[i].Value?.GetType() == typeof(UndertaleInstruction.Reference<UndertaleVariable>))
-                        Assert.AreSame(((UndertaleInstruction.Reference<UndertaleVariable>)code.Instructions[i].Value).Target, ((UndertaleInstruction.Reference<UndertaleVariable>)reasm[i].Value).Target, errMsg);
+                        Assert.True(Object.ReferenceEquals(((UndertaleInstruction.Reference<UndertaleVariable>)code.Instructions[i].Value).Target, ((UndertaleInstruction.Reference<UndertaleVariable>)reasm[i].Value).Target), errMsg);
                     else if (code.Instructions[i].Value?.GetType() == typeof(UndertaleResourceById<UndertaleString, UndertaleChunkSTRG>))
-                        Assert.AreSame(((UndertaleResourceById<UndertaleString, UndertaleChunkSTRG>)code.Instructions[i].Value).Resource, ((UndertaleResourceById<UndertaleString, UndertaleChunkSTRG>)reasm[i].Value).Resource, errMsg);
+                        Assert.True(Object.ReferenceEquals(((UndertaleResourceById<UndertaleString, UndertaleChunkSTRG>)code.Instructions[i].Value).Resource, ((UndertaleResourceById<UndertaleString, UndertaleChunkSTRG>)reasm[i].Value).Resource), errMsg);
                     else
-                        Assert.AreEqual(code.Instructions[i].Value, reasm[i].Value, errMsg);
+                        Assert.True(code.Instructions[i].Value == reasm[i].Value, errMsg);
                 }
             });
         }
     }
 
-    [TestClass]
     public class UndertaleLoadingTest : GameLoadingTestBase
     {
         public UndertaleLoadingTest() : base(GamePaths.UNDERTALE_PATH, GamePaths.UNDERTALE_MD5)
@@ -126,7 +131,6 @@ namespace UndertaleModTests
         }
     }
 
-    [TestClass]
     public class UndertaleSwitchLoadingTest : GameLoadingTestBase
     {
         public UndertaleSwitchLoadingTest() : base(GamePaths.UNDERTALE_SWITCH_PATH, GamePaths.UNDERTALE_SWITCH_MD5)
@@ -134,7 +138,6 @@ namespace UndertaleModTests
         }
     }
 
-    [TestClass]
     public class DeltaruneLoadingTest : GameLoadingTestBase
     {
         public DeltaruneLoadingTest() : base(GamePaths.DELTARUNE_PATH, GamePaths.DELTARUNE_MD5)
@@ -142,12 +145,9 @@ namespace UndertaleModTests
         }
     }
 
-
-
-    [TestClass]
     public class EmptyGameTest
     {
-        [TestMethod]
+        [Fact]
         public void CreateAndSaveEmptyGame()
         {
             UndertaleData data = UndertaleData.CreateNew();
@@ -157,5 +157,4 @@ namespace UndertaleModTests
             }
         }
     }
-
 }
