+++ UndertaleModLib.Tests/GamePaths.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace UndertaleModLib.Tests
+{
+    static class GamePaths
+    {
+        // TODO: Maybe these should be configuration variables?
+        //public static string UNDERTALE_PATH = @"C:\Program Files (x86)\Steam\steamapps\common\Undertale\data.win";
+        public static string UNDERTALE_PATH = @"D:\SteamLibrary\steamapps\common\Undertale\data.win";
+        public static string UNDERTALE_MD5 = "5903fc5cb042a728d4ad8ee9e949c6eb";
+        public static string UNDERTALE_SWITCH_PATH = @"..\..\..\Test\bin\Debug\switch\game.win";
+        public static string UNDERTALE_SWITCH_MD5 = "427520a97db28c87da4220abb3a334c1";
+        public static string DELTARUNE_PATH = @"C:\Program Files (x86)\SURVEY_PROGRAM\data.win";
+        public static string DELTARUNE_MD5 = "a88a2db3a68c714ca2b1ff57ac08a032";
+    }
+}
