+++ UndertaleModTool.sln
@@ -7,13 +7,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UndertaleModTool", "Underta
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UndertaleModLib", "UndertaleModLib\UndertaleModLib.csproj", "{4781ADAC-9469-42C4-B357-4A0A157D7FCD}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UndertaleModTests", "UndertaleModTests\UndertaleModTests.csproj", "{0B51CE42-C047-4B07-A8C9-88F4E3EA6DBD}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UndertaleModCli", "UndertaleModCli\UndertaleModCli.csproj", "{CC5FDC60-A955-4D0C-B59F-8BD451778237}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UndertaleModToolUpdater", "UndertaleModToolUpdater\UndertaleModToolUpdater.csproj", "{1E004B65-F3C8-4A62-B649-1465EA6A3B01}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UndertaleModToolUpdater", "UndertaleModToolUpdater\UndertaleModToolUpdater.csproj", "{1E004B65-F3C8-4A62-B649-1465EA6A3B01}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UndertaleModLibTests", "UndertaleModLibTests\UndertaleModLibTests.csproj", "{AAE3D4CA-0B6C-4928-AE86-066F5251F70A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UndertaleModLib.Tests", "UndertaleModLib.Tests\UndertaleModLib.Tests.csproj", "{84864F47-9788-415D-A51D-791BE2B6339B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -39,14 +37,6 @@ Global
 		{4781ADAC-9469-42C4-B357-4A0A157D7FCD}.Release|Any CPU.Build.0 = Release|Any CPU
 		{4781ADAC-9469-42C4-B357-4A0A157D7FCD}.Release|x64.ActiveCfg = Release|x64
 		{4781ADAC-9469-42C4-B357-4A0A157D7FCD}.Release|x64.Build.0 = Release|x64
-		{0B51CE42-C047-4B07-A8C9-88F4E3EA6DBD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0B51CE42-C047-4B07-A8C9-88F4E3EA6DBD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0B51CE42-C047-4B07-A8C9-88F4E3EA6DBD}.Debug|x64.ActiveCfg = Debug|x64
-		{0B51CE42-C047-4B07-A8C9-88F4E3EA6DBD}.Debug|x64.Build.0 = Debug|x64
-		{0B51CE42-C047-4B07-A8C9-88F4E3EA6DBD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0B51CE42-C047-4B07-A8C9-88F4E3EA6DBD}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0B51CE42-C047-4B07-A8C9-88F4E3EA6DBD}.Release|x64.ActiveCfg = Release|x64
-		{0B51CE42-C047-4B07-A8C9-88F4E3EA6DBD}.Release|x64.Build.0 = Release|x64
 		{CC5FDC60-A955-4D0C-B59F-8BD451778237}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{CC5FDC60-A955-4D0C-B59F-8BD451778237}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CC5FDC60-A955-4D0C-B59F-8BD451778237}.Debug|x64.ActiveCfg = Debug|x64
@@ -63,14 +53,14 @@ Global
 		{1E004B65-F3C8-4A62-B649-1465EA6A3B01}.Release|Any CPU.Build.0 = Release|Any CPU
 		{1E004B65-F3C8-4A62-B649-1465EA6A3B01}.Release|x64.ActiveCfg = Release|Any CPU
 		{1E004B65-F3C8-4A62-B649-1465EA6A3B01}.Release|x64.Build.0 = Release|Any CPU
-		{AAE3D4CA-0B6C-4928-AE86-066F5251F70A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AAE3D4CA-0B6C-4928-AE86-066F5251F70A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AAE3D4CA-0B6C-4928-AE86-066F5251F70A}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{AAE3D4CA-0B6C-4928-AE86-066F5251F70A}.Debug|x64.Build.0 = Debug|Any CPU
-		{AAE3D4CA-0B6C-4928-AE86-066F5251F70A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AAE3D4CA-0B6C-4928-AE86-066F5251F70A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{AAE3D4CA-0B6C-4928-AE86-066F5251F70A}.Release|x64.ActiveCfg = Release|Any CPU
-		{AAE3D4CA-0B6C-4928-AE86-066F5251F70A}.Release|x64.Build.0 = Release|Any CPU
+		{84864F47-9788-415D-A51D-791BE2B6339B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{84864F47-9788-415D-A51D-791BE2B6339B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{84864F47-9788-415D-A51D-791BE2B6339B}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{84864F47-9788-415D-A51D-791BE2B6339B}.Debug|x64.Build.0 = Debug|Any CPU
+		{84864F47-9788-415D-A51D-791BE2B6339B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{84864F47-9788-415D-A51D-791BE2B6339B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{84864F47-9788-415D-A51D-791BE2B6339B}.Release|x64.ActiveCfg = Release|Any CPU
+		{84864F47-9788-415D-A51D-791BE2B6339B}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
